Add an explicit CPU family AUTO
This allows user to explicitly set an automatic choosen CPU family. The old implemenation with the empty string
will cause endless churn loops, if the nodepool resource is created by another reconciler which has the empty string as desired value.
This is the case if the user uses ArgoCD or fluxCD.